+++ R/installAcidverse.R
@@ -1,0 +1,34 @@
+#' Install Acid Genomics packages
+#'
+#' @export
+#' @note Updated 2020-10-28.
+#'
+#' @examples
+#' ## > installAcidverse()
+installAcidverse <- function() {
+    install(
+        pkgs = c(
+            "AcidGSEA",                                 # RNASeq
+            "AcidGenerics",                             # Infrastructure
+            "AcidPlots",                                # Visualization
+            "AcidRoxygen",                              # Infrastructure
+            "AcidTest",                                 # Infrastructure
+            "Cellosaurus",                              # CancerData
+            "Chromium",                                 # SingleCell
+            "DESeqAnalysis",                            # RNASeq
+            "DepMapAnalysis",                           # CancerData
+            "EggNOG",                                   # Annotation
+            "PANTHER",                                  # Annotation
+            "WormBase",                                 # Annotation
+            "basejump",                                 # Infrastructure
+            "bb8",                                      # Infrastructure
+            "bcbioRNASeq",                              # RNASeq
+            "bcbioSingleCell",                          # SingleCell
+            "cBioPortalAnalysis",                       # CancerData
+            "goalie",                                   # Infrastructure
+            "pointillism",                              # SingleCell
+            "syntactic"                                 # Infrastructure
+        ),
+        reinstall = FALSE
+    )
+}